+++ src/core/utils/log.h
@@ -189,12 +189,12 @@ namespace allpix {
          * @brief Set the current event number from now on
          * @param event_num Event number to use
          */
-        static void setEventNum(unsigned int event_num);
+        static void setEventNum(uint64_t event_num);
         /**
          * @brief Get the current event numbers
          * @return Event number used
          */
-        static unsigned int getEventNum();
+        static uint64_t getEventNum();
 
     private:
         /**
@@ -225,7 +225,7 @@ namespace allpix {
 
         // Internal methods to store static values
         static std::string& get_section();
-        static unsigned int& get_event_num();
+        static uint64_t& get_event_num();
         static LogLevel& get_reporting_level();
         static LogFormat& get_format();
         static std::vector<std::ostream*>& get_streams();
